Let the service die when the device to read kbd codes does not exist
In case when TERES1 laptop is put to sleep and resumes, the service gets back online trying to read from gpio device, which does not exist anymore.

So the service does not work, but also is not stopped since all errors are ignore and as a "bonus" it actually takes 100% of the CPU core.

As a solution added exit condition in the main loop, when error ENODEV is received on read from the device. Also some headers were added to avoid warnings.